requirements.txt
requests>=2.28.0
icalendar>=5.0.0
flask>=2.0.0
python-dotenv>=1.0.0
